feat: update correct definition file
